transferred the first window code to main.c
